move shop, wallet module to organisation module
